requirements.txt
cassandra-driver==3.29.0
click==8.1.7
colorama==0.4.6
geomet==0.2.1.post1
six==1.16.0
